Moved analysis to panel
