+++ opal-ui/.eslintrc-auto-import.json
@@ -17,6 +17,7 @@
     "InjectionKey": true,
     "PropType": true,
     "Ref": true,
+    "TOKEN_TYPES": true,
     "VNode": true,
     "WritableComputedRef": true,
     "computed": true,
@@ -58,52 +59,48 @@
     "shallowReactive": true,
     "shallowReadonly": true,
     "shallowRef": true,
+    "stores": true,
     "toRaw": true,
     "toRef": true,
     "toRefs": true,
     "toValue": true,
     "triggerRef": true,
     "unref": true,
+    "useAppsStore": true,
     "useAttrs": true,
+    "useAuthStore": true,
+    "useAuthzStore": true,
+    "useCartStore": true,
+    "useCommandsStore": true,
     "useCssModule": true,
     "useCssVars": true,
+    "useDatashieldStore": true,
+    "useDatasourceStore": true,
+    "useFilesStore": true,
+    "useGroupsStore": true,
     "useI18n": true,
+    "useIdentifiersStore": true,
+    "useIdentityProvidersStore": true,
     "useLink": true,
+    "usePluginsStore": true,
+    "useProfileAclsStore": true,
+    "useProfilesStore": true,
+    "useProjectsStore": true,
+    "useRStore": true,
+    "useResourcesStore": true,
     "useRoute": true,
     "useRouter": true,
+    "useSearchStore": true,
     "useSlots": true,
+    "useSqlStore": true,
+    "useSystemStore": true,
+    "useTaxonomiesStore": true,
+    "useTokensStore": true,
+    "useTransientDatasourceStore": true,
+    "useUsersStore": true,
     "watch": true,
     "watchEffect": true,
     "watchPostEffect": true,
-    "watchSyncEffect": true,
-    "stores": true,
-    "useCounterStore": true,
-    "useAuthStore": true,
-    "useProjectsStore": true,
-    "useDatasourceStore": true,
-    "useFilesStore": true,
-    "usePluginsStore": true,
-    "useTransientDatasourceStore": true,
-    "useCommandsStore": true,
-    "useUsersStore": true,
-    "useUsersGroupsStore": true,
-    "useGroupsStore": true,
-    "useSystemStore": true,
-    "useRStore": true,
-    "useDatashieldStore": true,
-    "useAuthzStore": true,
-    "useProfilesStore": true,
-    "useProfileAclsStore": true,
-    "useTokensStore": true,
-    "useIdentityProvidersStore": true,
-    "TOKEN_TYPES": true,
-    "useTaxonomiesStore": true,
-    "useResourcesStore": true,
-    "useIdentifiersStore": true,
-    "useAppsStore": true,
-    "useSqlStore": true,
-    "useSearchStore": true,
-    "useCartStore": true,
-    "AnalysisPluginPackages": true
+    "watchSyncEffect": true
   }
 }
