+++ opal-ui/src/auto-imports.d.ts
@@ -12,7 +12,6 @@ declare global {
   const $ref: typeof import('vue/macros')['$ref']
   const $shallowRef: typeof import('vue/macros')['$shallowRef']
   const $toRef: typeof import('vue/macros')['$toRef']
-  const AnalysisPluginPackages: typeof import('./stores/plugins')['AnalysisPluginPackages']
   const EffectScope: typeof import('vue')['EffectScope']
   const TOKEN_TYPES: typeof import('./stores/tokens')['TOKEN_TYPES']
   const computed: typeof import('vue')['computed']
@@ -67,7 +66,6 @@ declare global {
   const useAuthzStore: typeof import('./stores/authz')['useAuthzStore']
   const useCartStore: typeof import('./stores/cart')['useCartStore']
   const useCommandsStore: typeof import('./stores/commands')['useCommandsStore']
-  const useCounterStore: typeof import('./stores/auth')['useCounterStore']
   const useCssModule: typeof import('vue')['useCssModule']
   const useCssVars: typeof import('vue')['useCssVars']
   const useDatashieldStore: typeof import('./stores/datashield')['useDatashieldStore']
@@ -93,7 +91,6 @@ declare global {
   const useTaxonomiesStore: typeof import('./stores/taxonomies')['useTaxonomiesStore']
   const useTokensStore: typeof import('./stores/tokens')['useTokensStore']
   const useTransientDatasourceStore: typeof import('./stores/transient-datasource')['useTransientDatasourceStore']
-  const useUsersGroupsStore: typeof import('./stores/users-groups')['useUsersGroupsStore']
   const useUsersStore: typeof import('./stores/users')['useUsersStore']
   const watch: typeof import('vue')['watch']
   const watchEffect: typeof import('vue')['watchEffect']
